Revert "Reorganize (#38)"
This reverts commit 56f4a3335de82fc31718da726d0db0bb9793df96.
